renames gem to HashComp
There is another gem defining a `HashDiff` constant in the global scope
(see [`hashdiff`](https://github.com/liufengyun/hashdiff)).

It might happen that both gems are included in a project, which leads to
conflicts as both libraries access the same constant.

A simple way to remove these conflicts is to rename one of the gem.

`hashdiff` is older than `hash_diff` (considering their initial commit)
and therefore the simple, dummy solution is to rename `hash_diff`.